refactor code and create composables useMap and useRouting
